+++ order.go
@@ -3,8 +3,8 @@ package sqlabble
 import (
 	"fmt"
 
+	"github.com/minodisk/sqlabble/internal/direction"
 	"github.com/minodisk/sqlabble/internal/generator"
-	"github.com/minodisk/sqlabble/internal/grammar/direction"
 )
 
 type order struct {
