+++ set_operation.go
@@ -2,7 +2,7 @@ package sqlabble
 
 import (
 	"github.com/minodisk/sqlabble/internal/generator"
-	"github.com/minodisk/sqlabble/internal/grammar/keyword"
+	"github.com/minodisk/sqlabble/internal/keyword"
 )
 
 type setOperation struct {
